requirements.txt
django==4.2.16
djangorestframework==3.15.2
python-dotenv==1.0.1
